remove tracing server
