+++ Makefile
@@ -1,6 +1,6 @@
-.PHONY: storage frontend client tracing-server config-gen clean
+.PHONY: storage frontend client config-gen clean
 
-all: storage frontend client tracing-server
+all: storage frontend client
 
 storage:
 	go build -o storage cmd/storage/main.go
@@ -11,8 +11,5 @@ frontend:
 client:
 	go build -o client cmd/client/main.go
 
-tracing-server:
-	go build -o tracing-server cmd/tracing-server/main.go
-
 clean:
-	rm storage frontend client tracing-server *".log" *"-Log.txt" 2> /dev/null || true
+	rm storage frontend client mem || true
