+++ cmd/frontend/main.go
@@ -1,12 +1,58 @@
 package main
 
 import (
+	"errors"
 	"fmt"
 	"log"
+	"net"
+	"net/http"
+	"net/rpc"
 
 	distkvs "example.org/cpsc416/a5"
-	"github.com/DistributedClocks/tracing"
+	"example.org/cpsc416/a5/kvslib"
 )
+
+type StorageAddr string
+
+type FrontEndConfig struct {
+	ClientAPIListenAddr  string
+	StorageAPIListenAddr string
+	Storage              StorageAddr
+}
+
+type FrontEndStorageFailed struct{}
+
+type FrontEndStorageStarted struct{}
+
+type FrontEndGet struct {
+	Key string
+}
+
+type FrontEndGetResult struct {
+	Key   string
+	Value *string
+	Err   bool
+}
+
+type StorageGet struct {
+	Key string
+}
+
+type StoragePut struct {
+	Key   string
+	Value string
+	delay int
+}
+
+type StorageGetResult struct {
+	Key   string
+	Value *string
+}
+
+type FrontEnd struct {
+	// state may go here
+	rpcClient *rpc.Client
+}
 
 func main() {
 	var config distkvs.FrontEndConfig
@@ -17,16 +63,82 @@ func main() {
 
 	fmt.Println(config)
 
-	tracer := tracing.NewTracer(tracing.TracerConfig{
-		ServerAddress:  config.TracerServerAddr,
-		TracerIdentity: "frontend",
-		Secret:         config.TracerSecret,
-	})
-
-	frontend := distkvs.FrontEnd{}
-	err = frontend.Start(config.ClientAPIListenAddr, config.StorageAPIListenAddr, 0, tracer)
+	frontend := FrontEnd{}
+	err = frontend.Start(config.ClientAPIListenAddr, config.StorageAPIListenAddr, 0)
 
 	if err != nil {
 		log.Fatal(err)
 	}
+}
+
+func (f *FrontEnd) Start(clientAPIListenAddr string, storageAPIListenAddr string, storageTimeout uint8) error {
+	// result := new(FrontEndGetResult)
+	result := new(FrontEnd)
+	rpc.Register(result)
+
+	// result2 := new(FrontEndPutResult)
+	// rpc.Register(result2)
+
+	rpc.HandleHTTP()
+
+	rpcClient, rpcErr := rpc.DialHTTP("tcp", storageAPIListenAddr)
+	if rpcErr != nil {
+		return errors.New("Cannot established connection with Storage server.")
+	}
+
+	result.rpcClient = rpcClient
+
+	l, e := net.Listen("tcp", clientAPIListenAddr)
+	if e != nil {
+		log.Fatal("listen error:", e)
+	}
+
+	err := http.Serve(l, nil)
+	if err != nil {
+		log.Fatal("listen error:", err)
+	}
+
+	return errors.New("Frontend Fail")
+}
+
+func (f *FrontEnd) HandleGet(args kvslib.KvslibGet, reply *kvslib.ResultStruct) error {
+	storageArgs := StorageGet{args.Key}
+	storageReply := new(StorageGetResult)
+	funcCall := f.rpcClient.Go("Storage.StorageGet", storageArgs, &storageReply, nil)
+	replyCall := <-funcCall.Done
+
+	// log.Print(replyCall.Error)
+	// log.Print(*storageReply.Value)
+
+	if replyCall.Error != nil {
+		return errors.New("FE to Strage fail")
+	}
+
+	ret := kvslib.ResultStruct{}
+	ret.Result = storageReply.Value
+	*reply = ret
+
+	log.Printf("OpId: %d Get value %s from %s", args.OpId, *reply.Result, args.Key)
+
+	return nil
+}
+
+func (f *FrontEnd) HandlePut(args kvslib.KvslibPut, reply *kvslib.ResultStruct) error {
+	storageArgs := StoragePut{args.Key, args.Value, 0}
+	storageReply := new(string)
+	funcCall := f.rpcClient.Go("Storage.StoragePut", storageArgs, &storageReply, nil)
+	replyCall := <-funcCall.Done
+
+	if replyCall.Error != nil {
+		return errors.New("FE to Strage fail")
+	}
+	// log.Print(*storageReply)
+
+	ret := kvslib.ResultStruct{}
+	ret.Result = storageReply
+	*reply = ret
+
+	log.Printf("OpId: %d Put value %s to %s", args.OpId, args.Value, args.Key)
+	// log.Print(*reply.Result)
+	return nil
 }
