+++ config/client_config.json
@@ -1,6 +1,4 @@
 {
   "ClientID": "client1",
-  "FrontEndAddr": ":8000",
-  "TracerServerAddr": ":6666",
-  "TracerSecret": ""
+  "FrontEndAddr": ":8000"
 }
