+++ config/storage_config.json
@@ -2,7 +2,5 @@
   "StorageID": "storage1",
   "StorageAdd": ":8011",
   "FrontEndAddr": ":8001",
-  "TracerServerAddr": ":6666",
-  "TracerSecret": "",
   "DiskPath" : "/home/boss/database-kvs/assignment-5/data"
 }
