--- config/tracing_server_config.json
@@ -1,6 +1,0 @@
-{
-    "ServerBind": ":6666",
-    "Secret": "",
-    "OutputFile": "trace_output.log",
-    "ShivizOutputFile": "shiviz_output.log"
-}
