+++ go.mod
@@ -1,5 +1,3 @@
 module example.org/cpsc416/a5
 
 go 1.14
-
-require github.com/DistributedClocks/tracing v0.0.0-20210325210840-4e33437ee7a9
